Reorganization
The checkin is a back and does not function. I does compile, however there are many logic areas. The usage of the cairo region object and intersection testing is incorrect. THe usage of cairo_region_create_rectangles is invalid. More research into the api usage has to be completed. The function directly affected is the iterate and plot function of the display context.