rendering efficiency
Rendering portions of the screen is much less due to the searching for on screen items. The system maintains two lists, onscreen and offscreen. The searching of intersection points is not optimized yet. A binary search Algorithm  may be used in conjunction with a sorted list or other possible solutions such as modulus banding.